requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pgvector==0.2.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Redis & Celery
redis==5.0.1
celery==5.3.6
flower==2.0.1

# AI & NLP
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5
spacy==3.7.2
tiktoken==0.5.2

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-magic==0.4.27

# AWS S3
boto3==1.34.34
botocore==1.34.34

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
email-validator==2.1.0

# Monitoring & Logging
python-json-logger==2.0.7
sentry-sdk==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Machine Learning
scikit-learn==1.4.0
joblib==1.3.2
